TAX_LAW_API/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain-openai>=0.2.0
langchain-pinecone>=0.2.0
pinecone-client>=5.0.1
python-multipart>=0.0.6
typing-extensions>=4.5.0
python-docx>=0.8.11  # For reading .docx files
requests>=2.31.0     # For downloading files from URLs 